chore: move config to utils
